Use non-numerical characters in LIS_FMT_*
In the future, when we want to support curves with lis::string as the
data type, we will need to embed the length of the string into the
format-string. To properly disambiguate between a string length and a
fmt-char in the format-string we change all the LIS_FMT_ characters to
be non-numeric.