Add support for lis::string in LIS Data Records
This commit adds support for LIS curves where the data-type is
lis::string.

lis::strings are a bit different from the other LIS79 data-types in that
it's not entirely self-contained. That is, the lis::string it not null
terminated nor does it embed it's length. For lis::strings in Data
Records the length is calculated from the reserved-size (and number of
samples) in the Spec Block [1]. The length is then embedded into the
format-string. E.g. the format-string for a 256 byte lis::string is
"A256".

[1] Note that this implicitly mean that curves using lis::string cannot
    have multiple entires pr sample.