Update fixed records tests
While using default file is ok in general, many fields in it have
default empty value (spaces repeated n times). As spaces can be seen as
default value of these fields, it decreases a bit fail-proofness of the
tests. While all the fields are already tested in c++, it would be nice
to have such a basic coverage in python as it provides user-facing code.

Hence tests are updated, especially given that files themselves already
exist.