Introduce ErrorHandler to load
Note that this commit changes current default LIS load behavior.
Before load of broken files almost always returned correctly. All the
errors were put in the log.
Now the same files will fail unless custom errorhandler is
injected into the load function.

ErrorHandler is not passed down to the logical file yet, but there are
plans to make it available at least for the curves.

Need to handle some warning messages might come in the future as well.

Note that commit leaks some filehandles. This is to be solved in the
future commit.