Sanity check DFSR before creating fmtstr and dtype
When creating both fmtstr and dtype there is a lot that can go wrong.
Hence the routines for crating these are heavily infested with sanity
checking of the DFSR. This seams like redundant work - while also
obscuring what these functions are actually tasked to do. Hence we
introduce a new function that does the necessary sanity checking of the
DFSR prior to the construction of fmtstr and dtype.

There are some additional notes to this commit:

 - The new method sanity-checks depth-recording-mode==1 and fast
   channels stuff, even if they are yet to implemented

 - It's stricter than before. Previously incorrect channels that where
   subject to suppression was allowed to be incorrect. This is no longer
   the case.

 - This commit does not remove the sanity checking from dfs_fmtstr as
   this function is subject to die soon anyway