Add support for DFSR with depth recording  mode 1
This commit introduces support for DFSRs that use the alternative depth
recording mode (1) where the index is only recorded once pr record.

In order to effectivly handle depth recording mode 1 (and future support
for Fast Channels) the definition of the fmtstr is completely re-done.
See the docs of dfsr_fmtstr.