Add index channel to test files
These test-files did not have an index channel. As of now that is not a
problem as we treat all channels equally. However, that is subject to
change when we introduce support for depth recording mode 1 and fast
channels. To avoid these test breaking in the future over un-related
index changes, we add a normal index channel to them.