Add a propper interface for Data Format Specs
These records are integral to the LIS standard and deserves a better
interface. DataFormatSpec is a wrapper for the more primitive C++ struct
lis::dfsr. It carries over the interface of lis::dfsr while introducing
a better interface for Entry Blocks. These are implemented as methods
with default values corresponding to what's defined in LIS79.