Drop support for "string" index in lis curves
Although nonsensical, the index channel can contain samples of type
lis::string. This happened to just work because we do not handle the
index in any special way at this point. However, that is subject to
change when support for depth recording mode 1 and fast channels is
introduced. In these cases an index with type string isn't just
nonsensical, but also impossible.

Support to avoid special casing and inconsistent behavior we drop
support all-together. And to be clear, we find it highly unlikely that
such files exist in real life.