Drop support for non-scalar index channel
Although nonsensical, the index channel can contain samples with
multiple entries. This happens to "just work" because we do not handle
the index in any special way at this point. However, that is subject to
change when support for depth recording mode 1 and fast channels is
introduced. In such cases the behavior of an index with non-scalar
sample values is undefined by the standard.

Support to avoid special casing and inconsistent behavior we drop
support all-together. And to be clear, we find it highly unlikely that
such files exist in real life.