Add support for LIS Fast Channels
This commit adds support for Fast Channels. This is the last piece of
the puzzle with respect to LIS curves. With exception of supporting
lis::mask as a valid curve-type, lis.curves is now feature complete.
Fast channels is independent of depth recording mode, and even works
with the special case of lis::string as datatype.

This commit do break the interface of lis.curves in that the parameter
`skip_fast` is now redundant and has been removed. A new parameter,
`sample_rate` is added to control which curves to read. Please refer to
the docs of lis.curves for a more in-dept description of the new
interface.

Co-authored-by: Alena Chaikouskaya <alec@equinor.com>