Do not assign value when entry size is 0
On reading entries we have 2 options when reading value: either go with
declared entry size or go with repcode size.

We used repcode size for all the fixed entries.
However, as per specification, terminator entry size 0 is a valid value.
Hence in that case we shouldn't incorrectly attempt to read any data.

Specification says nothing about 0 size for other types, but the same
logic is applied in implementation: declared size 0 passes validation
and entry value is not changed from default.