Make the default index mnemonic an attribute
In depth recording mode 1 the mnemonic of the index channel is not
explicitly defined in the file, nor in the LISS79 specification.
However, when depth recording mode is 1, the index is always referred to
as depth index by the spec. Hence 'DEPT' has been the hard-coded
default.  By adding it as an attribute on DataFormatSpec we have one
single source of truth, and it's customizable.