Add a Python ObjectStore for dlis metadata
Not caching metadata in python turns out to be really costly,
performance wise. Due to the interlinked object-to-object references,
even trivial user-patterns triggers multiple queries of the same
objects.

With the old store `dl::pool` that resides in C++, any object query
would trigger a trip to c++ where the object is looked for. The lookup
mechanism in dl::pool used python string-matchers, which essentially
meant a lot of trips back and forth between Python and c++ just to find
the correct object. When found, the object is returned to Python by
creating a Python wrapper (pybind11), which we then wrap in our own
Python classes (e.g. Channel, Frame, etc).

The new ObjectStore, is an python abstraction on top of `dl::pool`.
When caching the objects with ObjectStore, the aforementioned process is
reduces to a dict-loop for all successive queries. The caching can be
toggled on/off, and reset/cleared. It's on by default. When turned off,
the control flow is almost identical to the old behavior.