Pin scikit-build version in Appveyor
scikit-build version 0.12 breaks compilation on Appveyor.

It looks to be caused by [1] and specifically by [2].

Updates are stripping "Win64" from the provided compiler name.
scikit-build would attempt to use the 64-bit version of generator
instead.

Only that it looks like "Visual Studio 14 2015" (/bin/cl.exe) and
"Visual Studio 14 2015 Win64" (/bin/x86_amd64/cl.exe) are two totally
different compilers on Appveyor, located in different places and not
connected to each other. First is 32-bit and second is 64-bit.

So when Appveyor tries to call
-G "Visual Studio 14 2015" -A x64
it doesn't work, and 32-bit compiler "Visual Studio 14 2015" is used.
That in turn leads to clash when linking, as other libraries were
compiled with 64-bit compiler.

At the moment the simplest solution is just to pin the scikit-build
version to 0.11.0.

[1]
https://github.com/scikit-build/scikit-build/commit/3e3dc51dc04aafd0bb694139bbe5d2ebf0cac466
[2]
https://github.com/scikit-build/scikit-build/blob/3e3dc51dc04aafd0bb694139bbe5d2ebf0cac466/skbuild/platform_specifics/abstract.py#L292