Use rawlink for logo in README
Assuming that the logo file is always located in the same directory as
the README doesn't always hold. This is the case on e.g. PyPi, where
only the README itself read before rendering. Causing the logo to not
display. Setting the image to a rawlink rather then a relative path
solves that.