Run on job pr arch on each platform
Setting arch to auto on linux and macos in the strategy matrix means
that each invocation of cibuildwheel have to build wheels for multiple
arcs. Which means that extra care has to be taken to ensure that the
core is built for all archs.

Running one job per arch means we avoid all this nonsense.